Исправлены математические процедуры на работу с данныма типа double

+++ add.cpp
@@ -1,4 +1,4 @@
-int add(int a, int b)
+double add(double a, double b)
 {
     return a + b;
 }
