+++ mult.cpp
@@ -1,4 +1,4 @@
-int mult(int a, int b)
+double mult(double a, double b)
 {
     return a * b;
 }
